Homolog

+++ src/products/products.service.ts
@@ -2,6 +2,7 @@ import {
   forwardRef,
   Inject,
   Injectable,
+  InternalServerErrorException,
   NotFoundException,
 } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
@@ -22,7 +23,11 @@ export class ProductsService {
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return await this.productRepository.create(createProductDto, userId);
+    const product = await this.productRepository.create(
+      createProductDto,
+      userId,
+    );
+    return product;
   }
 
   async findAll(
@@ -46,7 +51,15 @@ export class ProductsService {
   }
 
   async findOne(id: number) {
-    return await this.productRepository.findOne(id);
+    try {
+      const product = await this.productRepository.findOne(id);
+      if (!product) {
+        throw new NotFoundException(`Product with ID ${id} not found`);
+      }
+      return product;
+    } catch (error) {
+      throw new NotFoundException('Error retrieving product', error);
+    }
   }
 
   async update(id: number, updateProductDto: UpdateProductDto) {
@@ -54,7 +67,18 @@ export class ProductsService {
   }
 
   async remove(id: number) {
-    return await this.productRepository.remove(id);
+    try {
+      const product = await this.productRepository.findOne(id);
+      if (!product) {
+        throw new NotFoundException(`Product with ID ${id} not found`);
+      }
+      await this.productRepository.remove(id);
+      return {
+        message: 'Product deleted successfully',
+      };
+    } catch (error) {
+      throw new InternalServerErrorException('Error deleting product', error);
+    }
   }
 
   async checkStock(productId: number, quantity: number) {
